exp to uint64_t impr broadcast (apply integral)

+++ src/base/math.h
@@ -285,27 +285,27 @@ float translate_to_percent(T from, T value, float maximum_percent)
 	}
 }
 
-constexpr inline unsigned long long computeExperience(unsigned Level)
+constexpr inline uint64_t computeExperience(unsigned Level)
 {
 	if(Level == 1)
 		return 18;
 	return Level * (static_cast<unsigned long long>(Level) - 1) * 24;
 }
 
-constexpr inline unsigned long long calculate_exp_gain(int factorCount, int baseLevel, int factorLevel)
+constexpr inline uint64_t calculate_exp_gain(int factorCount, int baseLevel, int factorLevel)
 {
 	int levelDifference = baseLevel - factorLevel;
 	double multiplier = (levelDifference >= 0)
 		? 1.0 + (levelDifference * 0.05)
 		: 1.0 / (1.0 + (std::abs(levelDifference) * 0.1));
-	unsigned long long baseExp = computeExperience(factorLevel) / factorCount;
-	unsigned long long experience = baseExp / multiplier;
-	unsigned long long minimumExperience = baseExp * 0.05;
-	return maximum((unsigned long long)1, experience, minimumExperience);
+	uint64_t baseExp = computeExperience(factorLevel) / factorCount;
+	uint64_t experience = baseExp / multiplier;
+	uint64_t minimumExperience = baseExp * 0.05;
+	return maximum((uint64_t)1, experience, minimumExperience);
 }
 
 
-constexpr inline unsigned long long calculate_gold_gain(int factorCount, int baseLevel, int factorLevel, bool randomBonus = false)
+constexpr inline uint64_t calculate_gold_gain(int factorCount, int baseLevel, int factorLevel, bool randomBonus = false)
 {
 	int levelDifference = baseLevel - factorLevel;
 	double multiplier = (levelDifference >= 0)
